Blacklist opam >= 2.2 due to bootstrap script failure
Starting with opam 2.2, the bootstrap script fails with:

> checking whether g++ really is a C++ compiler... yes
> configure: error: Dependencies missing. Use --with-vendored-deps or --disable-checks

Unlike the script in 2.1, which outputs a similar output but ends without an
error.

Note that the error only happens when a C++ compiler is detected. Otherwise,
the script will succeed, but the compiled opam will not have an external solver
and be mostly unusable.